% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SPA.R
\name{get_SPA}
\alias{get_SPA}
\title{get SPA (Suitability prevalence Area)}
\usage{
get_SPA(spatRaster, mean = "arithmetic")
}
\arguments{
\item{spatRaster}{a raster stack of environmental suitabilities along time}

\item{mean}{Type of mean to calculate. Could be arithmetic or geometric}
}
\value{
a raster with SPA values
}
\description{
get SPA (Suitability prevalence Area)
}
\examples{
rasterProjectionPresent <- terra::rast("data-raw/raster/rasterProjectionPresent.tif")
get_SPA(rasterProjectionPresent)
}
