% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Draw QQ plot}
\usage{
gg_qqplot(ps, ci = 0.95)
}
\arguments{
\item{ps}{A vector in which each element is an observed p value.}

\item{ci}{A parameter to control tolerance.}
}
\value{
A QQ plot of p values.
}
\description{
Draw QQ plot for p values.
}
\examples{
gg_qqplot(stats::runif(100, min = 0, max = 1))
}
