% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_simulation.R
\name{logistic_simulation}
\alias{logistic_simulation}
\title{Simulations in logistic regression}
\usage{
logistic_simulation(
  cases,
  controls,
  maf,
  X1prob1_case = 0.7260159,
  X1prob1_control = 0.4974979,
  samplerX2_given_Y0 = arscpp::ars(f = Log_Dens_X2_given_Y0, f_prime =
    Log_Dens_X2_given_Y0_prime, xlb = -Inf, xrb = Inf, x = c(-2, 0, 2), beta0 = -5.6,
    prevalence = 0.0109492),
  samplerX2_given_Y1 = arscpp::ars(f = Log_Dens_X2_given_Y1, f_prime =
    Log_Dens_X2_given_Y1_prime, xlb = -Inf, xrb = Inf, x = c(-2, 0, 2), beta0 = -5.6,
    prevalence = 0.0109492)
)
}
\arguments{
\item{samplerX2_given_Y1}{}
}
\value{
A list.
}
\description{
Numerical simulations in logistic regression model.
}
\references{
Dey, Rounak, et al. "A fast and accurate algorithm to test for binary phenotypes and its application to PheWAS." The American Journal of Human Genetics 101.1 (2017): 37-49.
}
