% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_spa_fast.R
\name{logistic_spa_fast}
\alias{logistic_spa_fast}
\title{Fast SPA in logistic regression}
\usage{
logistic_spa_fast(
  genos,
  pheno,
  cov,
  obj.null,
  minmac = 5,
  cutoff = 2,
  log.p = FALSE
)
}
\arguments{
\item{genos}{A vector or matrix containing the genotypes. If matrix is provided then rows should correspond to SNPs and columns should correspond to subjects. Optional, but needed if \code{obj.null} is missing.}

\item{pheno}{A vector containing the phenotypes. Optional, but needed if \code{obj.null} is missing.}

\item{cov}{A matrix or data frame containing the covariates. Optional, but needed if \code{obj.null} is missing.}

\item{obj.null}{An object of class "\code{logistic_null}". (Optional)}

\item{minmac}{An integer denoting the minimum minor allele count threshold to run SPA test. Default value is 5.}

\item{cutoff}{An integer denoting the standard deviation cutoff to be used. If the test statistic lies within the standard deviation cutoff of the mean, p-value based on traditional score test is returned. Default value is 2.}

\item{log.p}{A Boolean value determines whether to return natural log-transformed p-values. Default value is \code{FALSE}.}
}
\value{
A list of p values.
}
\description{
Fast saddlepoint approximation method in logistic regression model.
}
\examples{
# See vignettes
}
\references{
Zhou, Wei, et al. "Efficiently controlling for case-control imbalance and sample relatedness in large-scale genetic association studies." Nature genetics 50.9 (2018): 1335-1341.
}
