% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinom_spa_basic.R
\name{negbinom_spa}
\alias{negbinom_spa}
\title{Basic SPA in NB generalized linear model}
\usage{
negbinom_spa(
  perturbation,
  gene,
  cov,
  obj.null,
  minperturb = 5,
  cutoff = 2,
  log.p = FALSE
)
}
\arguments{
\item{perturbation}{A vector or matrix containing the perturbation indicators. Optional, but needed if \code{obj.null} is missing.}

\item{gene}{A vector containing the gene expressions. Optional, but needed if \code{obj.null} is missing.}

\item{cov}{A matrix or data frame containing the covariates. Optional, but needed if \code{obj.null} is missing.}

\item{obj.null}{An object of class "\code{logistic_null}". (Optional)}

\item{minperturb}{An integer denoting the minimum perturbation number. Default value is 5.}

\item{cutoff}{An integer denoting the standard deviation cutoff to be used. If the test statistic lies within the standard deviation cutoff of the mean, p-value based on traditional score test is returned. Default value is 2.}

\item{log.p}{A Boolean value determines whether to return natural log-transformed p-values. Default value is \code{FALSE}.}
}
\value{
A list of p values.
}
\description{
Basic saddlepoint approximation method in NB generalized linear model.
}
\examples{
# See vignettes
}
