% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_spa_basic.R
\name{logistic_spa}
\alias{logistic_spa}
\title{Basic SPA in logistic regression}
\usage{
logistic_spa(
  genos,
  pheno,
  cov,
  obj.null,
  minmac = 5,
  cutoff = 2,
  log.p = FALSE
)
}
\arguments{
\item{log.p}{}
}
\value{
A list.
}
\description{
Basic saddlepoint approximation method in logistic regression model.
}
\references{
Dey, Rounak, et al. "A fast and accurate algorithm to test for binary phenotypes and its application to PheWAS." The American Journal of Human Genetics 101.1 (2017): 37-49.
}
